Generate dense matrices (via spectral decomposition)

Generate p-dimensional matrices so that its inverse is dense. The matrix will be generated so its first 'num' eigen values are 1000 and the remaining are 1. The orthogonal basis is generated via QR decomposition of

denseQR(p = 8, num = 5, n = NULL)

Arguments

p
desired dimension

num
number of 'large' eigen values. Note num must be less than p

n
option to generate n observations from covariance matrix S

Examples

denseQR(p = 10, num = 10)

#> $Omega
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  1.000000e-03  7.105427e-20 -1.492140e-19 -2.153833e-20 -1.484923e-20
#>  [2,]  1.421085e-20  1.000000e-03  2.842171e-20 -1.332268e-19 -5.562217e-20
#>  [3,] -1.492140e-19 -2.842171e-20  1.000000e-03 -4.307665e-20  3.347322e-20
#>  [4,] -2.153833e-20 -1.332268e-19 -4.352074e-20  1.000000e-03  4.065845e-20
#>  [5,] -1.840195e-20 -6.977752e-20  8.987255e-20 -1.973768e-20  1.000000e-03
#>  [6,] -2.042810e-20  3.552714e-20  1.314504e-19  1.495470e-19  1.021822e-19
#>  [7,]  5.684342e-20 -1.705303e-19  9.237056e-20  7.999157e-20 -2.629841e-21
#>  [8,] -7.815970e-20  9.947598e-20 -2.131628e-19 -2.180478e-19 -6.411538e-20
#>  [9,] -4.085621e-20 -9.592327e-20 -3.907985e-20  2.192690e-20  5.507400e-20
#> [10,] -8.526513e-20 -5.684342e-20 -1.136868e-19 -1.509903e-19  8.398837e-20
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -7.105427e-20  5.506706e-20 -9.947598e-20 -4.263256e-20 -8.526513e-20
#>  [2,]  3.552714e-20 -1.136868e-19  1.278977e-19 -9.592327e-20 -1.136868e-19
#>  [3,]  1.314504e-19  9.237056e-20 -1.989520e-19 -1.065814e-20 -1.705303e-19
#>  [4,]  9.275913e-20  1.296740e-19 -2.178258e-19  5.034861e-20 -1.652012e-19
#>  [5,]  1.155048e-19 -2.602085e-21 -9.076073e-20  1.119174e-19  6.977752e-20
#>  [6,]  1.000000e-03  1.243450e-20  9.059420e-20  1.141309e-19 -3.552714e-21
#>  [7,] -1.598721e-20  1.000000e-03 -6.779273e-35  3.419487e-20  1.421085e-20
#>  [8,]  8.881784e-20  2.842171e-20  1.000000e-03 -6.927792e-20  8.526513e-20
#>  [9,]  1.429967e-19  6.350476e-20 -3.907985e-20  1.000000e-03  7.105427e-20
#> [10,] -3.552714e-21  1.421085e-20  8.526513e-20  7.105427e-20  1.000000e-03
#> 
#> $S
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  1.000000e+03 -7.105427e-14  1.492140e-13  2.153833e-14  1.484923e-14
#>  [2,] -1.421085e-14  1.000000e+03 -2.842171e-14  1.332268e-13  5.562217e-14
#>  [3,]  1.492140e-13  2.842171e-14  1.000000e+03  4.307665e-14 -3.347322e-14
#>  [4,]  2.153833e-14  1.332268e-13  4.352074e-14  1.000000e+03 -4.065845e-14
#>  [5,]  1.840195e-14  6.977752e-14 -8.987255e-14  1.973768e-14  1.000000e+03
#>  [6,]  2.042810e-14 -3.552714e-14 -1.314504e-13 -1.495470e-13 -1.021822e-13
#>  [7,] -5.684342e-14  1.705303e-13 -9.237056e-14 -7.999157e-14  2.629841e-15
#>  [8,]  7.815970e-14 -9.947598e-14  2.131628e-13  2.180478e-13  6.411538e-14
#>  [9,]  4.085621e-14  9.592327e-14  3.907985e-14 -2.192690e-14 -5.507400e-14
#> [10,]  8.526513e-14  5.684342e-14  1.136868e-13  1.509903e-13 -8.398837e-14
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,]  7.105427e-14 -5.506706e-14  9.947598e-14  4.263256e-14  8.526513e-14
#>  [2,] -3.552714e-14  1.136868e-13 -1.278977e-13  9.592327e-14  1.136868e-13
#>  [3,] -1.314504e-13 -9.237056e-14  1.989520e-13  1.065814e-14  1.705303e-13
#>  [4,] -9.275913e-14 -1.296740e-13  2.178258e-13 -5.034861e-14  1.652012e-13
#>  [5,] -1.155048e-13  2.602085e-15  9.076073e-14 -1.119174e-13 -6.977752e-14
#>  [6,]  1.000000e+03 -1.243450e-14 -9.059420e-14 -1.141309e-13  3.552714e-15
#>  [7,]  1.598721e-14  1.000000e+03  0.000000e+00 -3.419487e-14 -1.421085e-14
#>  [8,] -8.881784e-14 -2.842171e-14  1.000000e+03  6.927792e-14 -8.526513e-14
#>  [9,] -1.429967e-13 -6.350476e-14  3.907985e-14  1.000000e+03 -7.105427e-14
#> [10,]  3.552714e-15 -1.421085e-14 -8.526513e-14 -7.105427e-14  1.000000e+03
#>