insert_second_header()

Function to add a second header row

insert_second_header(
  wb,
  sheetname,
  data_start_row,
  group_names,
  grouplines,
  data
)

Arguments

wb

workbook object to add new worksheet to.

sheetname

Names of the worksheet in output file. Note that this name will be truncated to 31 characters, must be unique, and cannot contain some special characters (namely the following: /, \, ?, *, :, [, ]).

data_start_row

Row index for first row with data

group_names

A vector of names for groups to be displayed in a secondary header. Should be of the same length as grouplines, and cannot be used unless these are set. Defaults to NA, meaning no secondary header is created.

grouplines

Can be used to visually group variables. Values should either correspond to numeric column indices or column names, denoting the first variable in a group. Defaults to NA, meaning no lines are added.

data

data to be included.