insert_worksheet()

Function to add formatted worksheets to an existing .xlsx-workbook.

Usage

insert_worksheet(
  data,
  workbook,
  sheetname = "data",
  title = "Title",
  source = "statzh",
  metadata = NA,
  logo = "statzh",
  grouplines = FALSE,
  contactdetails = "statzh",
  author = "user"
)

Arguments

data

data to be included.

workbook

workbook object to add new worksheet to.

sheetname

name of the sheet tab.

title

title to be put above the data.

source

source of the data. Defaults to "statzh".

metadata

metadata information to be included. Defaults to NA.

logo

path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"

grouplines

defaults to FALSE. Can be used to separate grouped variables visually.

contactdetails

contact details of the data publisher. Defaults to "statzh".

author

defaults to the last two letters (initials) or numbers of the internal user name.

Examples

# Generation of a spreadsheet
wb <- openxlsx::createWorkbook("hello")

insert_worksheet(data = mtcars[c(1:10),], workbook = wb, title = "mtcars", sheetname = "carb")