splitXLSX()

Function to export data from R as formatted .xlsx-file and spread them over several worksheets based on a grouping variable (e.g., year).

splitXLSX(
  data,
  file,
  sheetvar,
  title = "Titel",
  source = "statzh",
  metadata = NA,
  logo = "statzh",
  grouplines = FALSE,
  contactdetails = "statzh",
  author = "user"
)

Arguments

data
data to be exported.

file
file name of the xlsx-file. The extension ".xlsx" is added automatically.

sheetvar
name of the variable used to split the data and spread them over several sheets.

title
title to be put above the data in the worksheet. the sheetvar subcategory is added in brackets.

source
source of the data. Defaults to "statzh".

metadata
metadata-information to be included. Defaults to NA.

logo
path of the file to be included as logo (png / jpeg / svg). Defaults to "statzh"

grouplines
columns to be separated visually by vertical lines.

contactdetails
contactdetails of the data publisher. Defaults to "statzh".

author
defaults to last two letters (initials) or numbers of the user name.

Examples

dat <- mtcars

splitXLSX(data = dat,
         title = "Motor trend car road tests",
         file = "motor_trend_car_road_tests", # '.xlsx' automatically added
         sheetvar = cyl,
         source = "Source: Henderson and Velleman (1981),
         Building multiple regression models interactively.
         Biometrics, 37, 391–411.",
         metadata = c("The data was extracted from the 1974
         Motor Trend US magazine and comprises fuel consumption and
         10 aspects of automobile design and performance for
         32 automobiles (1973–74 models)."),
         contactdetails = "statzh",
         grouplines = FALSE,
         logo = "statzh",
         author = "user")

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL

#> Warning: converting NULL pointer to R NULL