datasetsXLSX()

Function to export several datasets and/or figures from R to an .xlsx-file. The function creates an overview sheet and separate sheets for each dataset/figure.

datasetsXLSX(
  file,
  maintitle,
  datasets,
  widths = NULL,
  heights = NULL,
  startrows = NULL,
  startcols = NULL,
  sheetnames,
  titles,
  logo = "statzh",
  titlesource = "statzh",
  sources = "statzh",
  metadata1 = NA,
  auftrag_id = NULL,
  contact = "statzh",
  homepage = "statzh",
  openinghours = "statzh"
)

Arguments

file
file name of the spreadsheet. The extension ".xlsx" is added automatically.

maintitle
Title to be put on the first (overview) sheet.

datasets
datasets or plots to be included.

widths
width of figure in inch (1 inch = 2.54 cm). See details.

heights
height of figure in inch (1 inch = 2.54 cm). See details.

startrows
row where upper left corner of figure should be placed. See details.

startcols
column where upper left corner of figure should be placed. See details.

sheetnames
names of the sheet tabs.

titles
titles of the different sheets.

logo
file path to the logo to be included in the index-sheet. Can be "statzh" or "zh". Defaults to "statzh".

titlesource
source to be mentioned on the title sheet beneath the title

sources
source of the data. Defaults to "statzh".

metadata1
metadata information to be included. Defaults to NA.

auftrag_id
order number.

contact
contact information on the title sheet. Defaults to "statzh"

homepage
web address to be put on the title sheet. Default to "statzh"

openinghours
openinghours written on the title sheet. Defaults to Data Shop

Details

When including figures, the heights and widths need to be specified as a vector. For example, say you have two datasets and one figure that you would like to export. widths = c(0,0,5) then suggests that the figure will be 5 inches wide and placed in the third (and last) sheet of the file. The same logic applies to the specification of `startrows` and `startcols`. To include a ggplot2-plot, put a print statement around the ggplot2-object: list(..., ..., print(plot)).

Examples

# \donttest{
if (FALSE) {

# Example with two datasets and no figure
dat1 <- mtcars
dat2 <- PlantGrowth

datasetsXLSX(file="twoDatasets", # '.xlsx' wird automatisch hinzugefügt
            maintitle = "Autos und Pflanzen",
            datasets = list(dat1, dat2),
            logo = "statzh",
            sheetnames = c("Autos","Blumen"),
            titles = c("mtcars-Datensatz","PlantGrowth-Datensatz"),
            sources = c("Source: Henderson and Velleman (1981).
            Building multiple regression models interactively. Biometrics, 37, 391–411.",
                        "Dobson, A. J. (1983) An Introduction to Statistical
                        Modelling. London: Chapman and Hall."),
            metadata1 = c("Bemerkungen zum mtcars-Datensatz: x",
                          "Bemerkungen zum PlantGrowth-Datensatz: x"),
            auftrag_id="A2021_0000")

# Example with two datasets and one figure

dat1 <- mtcars
dat2 <- PlantGrowth
fig <- hist(mtcars$disp)

datasetsXLSX(file="twoDatasetsandFigure",
            maintitle = "Autos und Pflanzen", # '.xlsx' wird automatisch hinzugefügt
            datasets = list(dat1, dat2, fig),
            widths = c(0,0,5),
            heights = c(0,0,5),
            startrows = c(0,0,3),
            startcols = c(0,0,3),
            sheetnames = c("Autos","Blumen", "Histogramm"),
            titles = c("mtcars-Datensatz","PlantGrowth-Datensatz", "Histogramm"),
            sources = c("Source: Henderson and Velleman (1981).
            Building multiple regression models interactively. Biometrics, 37, 391–411.",
                        "Source: Dobson, A. J. (1983) An Introduction to
                        Statistical Modelling. London: Chapman and Hall."),
            metadata1 = c("Bemerkungen zum mtcars-Datensatz: x",
                          "Bemerkungen zum PlantGrowth-Datensatz: x"),
            auftrag_id="A2021_0000")
}
# }