datasetsXLSX

Function to create formatted XLSX files automatically

datasetsXLSX(
  file,
  maintitle,
  datasets,
  widths,
  heights,
  startrows,
  startcols,
  sheetnames,
  titles,
  logo = "statzh",
  sources = NULL,
  metadata1 = "",
  auftrag_id
)

Arguments

file
filename of the XLSX file. No Default.

maintitle
main title of first sheet "Inhalt".

datasets
datasets (dataframes or images) to be included in the XLSX-table.

widths
width of figure.

heights
height of figure.

startrows
row coordinate of upper left corner of figure.

startcols
column coordinate of upper left corner of figure.

sheetnames
name of the sheet tab.

titles
title of the table in the worksheet, defaults to "Titel" + the value of the variable used to split the dataset across sheets.

logo
filepath to the logo to be included in the index-sheet (default: Logo of the statistical office ZH)

sources
source of the data. Defaults to "statzh".

metadata1
metadata-information to be included. Defaults to NA.

auftrag_id
order number.

Examples

# \donttest{
# example with image / plot
# create map
if (FALSE) {
plot <- plot(mtcars$mpg)

datasetsXLSX(file="datasetsXLSX",
            maintitle = "nice datasets",
            datasets = list(head(mtcars),tail(mtcars),plot),
            sheetnames = c("data1","data2","map"),
            widths = c(0,0,5),
            heights = c(0,0,2.5),
            startrows = c(0,0,10),
            startcols = c(0,0,8),
            titles = c("Title","Title", "Plot"),
            sources = c("Quelle: STATENT", "Quelle: Strukturerhebung", "Quelle: Strukturerhebung"),
            metadata1 = c("Bemerkungen: bla", "Bemerkungen: y", "Bemerkungen: x"),
            auftrag_id="A2020_0200"
)

}
# example without image

datasetsXLSX(file="datasetsXLSX2",
             maintitle = "nice datasets",
             datasets = list(head(mtcars),tail(mtcars)),
             sheetnames = c("data1","data2"),
             titles = c("Title","Title"),
             sources = c("Quelle: STATENT", "Quelle: Strukturerhebung"),
             metadata1 = c("Bemerkungen: y", "Bemerkungen: x"),
             auftrag_id="AS2020_01"
             )

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: the condition has length > 1 and only the first element will be used

#> Warning: argument 'pattern' has length > 1 and only the first element will be used

#> Error in openxlsx::saveWorkbook(wb, paste(file, ".xlsx", sep = "")): File already exists!

# }