Generate dense matrices (via spectral decomposition)

Generate p-dimensional matrices so that its inverse is dense. The matrix will be generated so its first 'num' eigen values are 1000 and the remaining are 1. The orthogonal basis is generated via QR decomposition of

denseQR(p = 8, num = 5, n = NULL)

Arguments

p
desired dimension

num
number of 'large' eigen values. Note num must be less than p

n
option to generate n observations from covariance matrix S

Examples

denseQR(p = 10, num = 10)

#> $Omega
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  1.000000e-03 -9.009746e-20 -1.775087e-19 -5.094691e-20  3.797512e-20
#>  [2,] -5.619346e-20  1.000000e-03 -1.718992e-19 -5.223156e-21 -1.364408e-19
#>  [3,] -1.753421e-19 -1.808439e-19  1.000000e-03  2.690234e-20  1.122728e-19
#>  [4,] -4.278691e-20 -6.887481e-21  3.706987e-20  1.000000e-03  6.956295e-20
#>  [5,]  2.396316e-20 -1.507151e-19  1.124211e-19 -4.427801e-20  1.000000e-03
#>  [6,] -3.776440e-20 -1.098908e-19  1.893685e-19  1.189524e-19  2.943781e-19
#>  [7,]  2.891273e-20  4.097053e-20  7.147228e-20  6.755372e-20  1.627273e-21
#>  [8,] -5.771525e-20  1.485443e-19 -1.473958e-19 -1.375443e-19 -1.870193e-19
#>  [9,] -2.220830e-20  5.274968e-20  9.871813e-20  6.309109e-20  2.286281e-20
#> [10,]  1.022251e-19  2.086867e-20 -1.682423e-19 -2.000679e-20 -5.078412e-20
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,] -1.121806e-19  2.810948e-20 -5.663990e-20 -6.273476e-20  8.134849e-20
#>  [2,] -1.109852e-19  1.206047e-20  1.429966e-19  6.565664e-20  2.533708e-20
#>  [3,]  1.643783e-19  6.286331e-20 -1.478337e-19  1.236098e-19 -1.205085e-19
#>  [4,]  1.083015e-19  9.538618e-20 -1.421729e-19  7.042506e-20 -1.192233e-20
#>  [5,]  2.656255e-19 -1.259086e-20 -2.154301e-19  8.778578e-20 -7.278372e-20
#>  [6,]  1.000000e-03 -1.910343e-20 -5.950349e-20  1.646740e-19 -5.362177e-20
#>  [7,] -3.260344e-20  1.000000e-03  5.023596e-20 -3.231892e-20  8.890751e-21
#>  [8,] -5.060010e-20  8.134263e-20  1.000000e-03 -1.537786e-19  7.930495e-20
#>  [9,]  1.829021e-19 -3.387486e-20 -1.803460e-19  1.000000e-03 -5.442222e-20
#> [10,]  1.270469e-20  2.185313e-20  1.018579e-19 -9.126595e-20  1.000000e-03
#> 
#> $S
#>                [,1]          [,2]          [,3]          [,4]          [,5]
#>  [1,]  1.000000e+03  9.009746e-14  1.775087e-13  5.094691e-14 -3.797512e-14
#>  [2,]  5.619346e-14  1.000000e+03  1.718992e-13  5.223156e-15  1.364408e-13
#>  [3,]  1.753421e-13  1.808439e-13  1.000000e+03 -2.690234e-14 -1.122728e-13
#>  [4,]  4.278691e-14  6.887481e-15 -3.706987e-14  1.000000e+03 -6.956295e-14
#>  [5,] -2.396316e-14  1.507151e-13 -1.124211e-13  4.427801e-14  1.000000e+03
#>  [6,]  3.776440e-14  1.098908e-13 -1.893685e-13 -1.189524e-13 -2.943781e-13
#>  [7,] -2.891273e-14 -4.097053e-14 -7.147228e-14 -6.755372e-14 -1.627273e-15
#>  [8,]  5.771525e-14 -1.485443e-13  1.473958e-13  1.375443e-13  1.870193e-13
#>  [9,]  2.220830e-14 -5.274968e-14 -9.871813e-14 -6.309109e-14 -2.286281e-14
#> [10,] -1.022251e-13 -2.086867e-14  1.682423e-13  2.000679e-14  5.078412e-14
#>                [,6]          [,7]          [,8]          [,9]         [,10]
#>  [1,]  1.121806e-13 -2.810948e-14  5.663990e-14  6.273476e-14 -8.134849e-14
#>  [2,]  1.109852e-13 -1.206047e-14 -1.429966e-13 -6.565664e-14 -2.533708e-14
#>  [3,] -1.643783e-13 -6.286331e-14  1.478337e-13 -1.236098e-13  1.205085e-13
#>  [4,] -1.083015e-13 -9.538618e-14  1.421729e-13 -7.042506e-14  1.192233e-14
#>  [5,] -2.656255e-13  1.259086e-14  2.154301e-13 -8.778578e-14  7.278372e-14
#>  [6,]  1.000000e+03  1.910343e-14  5.950349e-14 -1.646740e-13  5.362177e-14
#>  [7,]  3.260344e-14  1.000000e+03 -5.023596e-14  3.231892e-14 -8.890751e-15
#>  [8,]  5.060010e-14 -8.134263e-14  1.000000e+03  1.537786e-13 -7.930495e-14
#>  [9,] -1.829021e-13  3.387486e-14  1.803460e-13  1.000000e+03  5.442222e-14
#> [10,] -1.270469e-14 -2.185313e-14 -1.018579e-13  9.126595e-14  1.000000e+03
#>